Rename to reflect-cpp and use Conan 2.x
