+++ recipes/reflect-cpp/all/test_package/CMakeLists.txt
@@ -1,12 +1,9 @@
-cmake_minimum_required(VERSION 3.12)
+cmake_minimum_required(VERSION 3.15)
 project(test_package LANGUAGES CXX)
 
-find_package(reflect-cpp REQUIRED CONFIG)
+find_package(reflectcpp REQUIRED CONFIG)
 
 add_executable(${PROJECT_NAME} test_package.cpp)
-target_link_libraries(${PROJECT_NAME} PRIVATE reflect-cpp::reflect-cpp)
+target_link_libraries(${PROJECT_NAME} reflectcpp::reflectcpp)
 target_compile_features(${PROJECT_NAME} PRIVATE cxx_std_20)
 
-if(CONAN_TEST_WITH_MSGPACK)
-  target_compile_definitions(${PROJECT_NAME} PRIVATE CONAN_TEST_WITH_MSGPACK)
-endif()
