+++ recipes/reflect-cpp/all/test_package/test_package.cpp
@@ -1,29 +1,15 @@
-#include <string>
+#include <cstdlib>
 #include <iostream>
-#include <rfl.hpp>
+#include <string>
 
-#if defined(CONAN_TEST_WITH_MSGPACK)
-  #include <rfl/msgpack.hpp>
-#endif
-
-struct TestStruct {
-    int x;
-    std::string name;
-};
+#include "rfl.hpp"
+#include "rfl/Generic.hpp"
+#include "rfl/json.hpp"
 
 int main(void) {
-    for (const auto& f : rfl::fields<TestStruct>()) {
-        (void) f.name();
-        (void) f.type();
-    }
-
-#if defined(CONAN_TEST_WITH_MSGPACK)
-    const auto test = TestStruct{.x = 15, .name = "test_package"};
-    std::cout << "msgpack test: ";
-    rfl::msgpack::write(test, std::cout) << std::endl;
-#endif
-
-    std::cout << "reflect-cpp test successful\n";
-
-    return 0;
+    auto person = rfl::Generic::Object();
+    person["first_name"] = "John";
+    person["last_name"] = "Doe";
+    rfl::json::write(person, std::cout) << std::endl;
+    return EXIT_SUCCESS;
 }
