+++ recipes/reflect-cpp/config.yml
@@ -1,9 +1,3 @@
 versions:
-  "0.11.1":
-    folder: all
-  "0.11.0":
-    folder: all
-  "0.10.0":
-    folder: all
-  "0.6.0":
+  "0.16.0":
     folder: all
