--- recipes/reflectcpp/all/test_package/CMakeLists.txt
@@ -1,9 +1,0 @@
-cmake_minimum_required(VERSION 3.15)
-project(test_package LANGUAGES CXX)
-
-find_package(reflectcpp REQUIRED CONFIG)
-
-add_executable(${PROJECT_NAME} test_package.cpp)
-target_link_libraries(${PROJECT_NAME} reflectcpp::reflectcpp)
-target_compile_features(${PROJECT_NAME} PRIVATE cxx_std_20)
-
