--- recipes/reflectcpp/config.yml
@@ -1,3 +1,0 @@
-versions:
-  "0.16.0":
-    folder: all
